requirements.txt
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
mutagen==1.47.0
pyaes==1.6.1
pycryptodomex==3.19.0
Pyrogram==2.0.106
pyromod==3.1.6
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
TgCrypto==1.2.5
Unidecode==1.3.7
urllib3==2.1.0
websockets==12.0
youtube-dl==2021.12.17
yt-dlp==2023.11.16
